+++ Lilu/Headers/kern_mach.hpp
@@ -30,6 +30,7 @@ class MachInfo {
 	mach_vm_address_t prelink_vmaddr {0};    // prelink text base vm address (for kexts this is their actual slide)
 	uint32_t file_buf_size {0};              // read file data size
 	uint8_t *linkedit_buf {nullptr};         // pointer to __LINKEDIT buffer containing symbols to solve
+	bool linkedit_buf_ro {false};            // linkedit_buf is read-only (not copy).
 	uint64_t linkedit_fileoff {0};           // __LINKEDIT file offset so we can read
 	uint64_t linkedit_size {0};
 	uint32_t symboltable_fileoff {0};        // file offset to symbol table - used to position inside the __LINKEDIT buffer
@@ -41,6 +42,7 @@ class MachInfo {
 	bool kaslr_slide_set {false};            // kaslr can be null, used for disambiguation
 	bool allow_decompress {true};            // allows mach decompression
 	bool prelink_slid {false};               // assume kaslr-slid kext addresses
+	bool kernel_collection {false};          // kernel collection (11.0+)
 	uint64_t self_uuid[2] {};                // saved uuid of the loaded kext or kernel
 
 	/**
@@ -148,6 +150,13 @@ class MachInfo {
 	 */
 	kern_return_t initFromFileSystem(const char * const paths[], size_t num);
 
+	/**
+	 *  Resolve mach data in the kernel via memory access
+	 *
+	 *  @return KERN_SUCCESS if loaded
+	 */
+	kern_return_t initFromMemory();
+
 public:
 
 	/**
@@ -192,6 +201,16 @@ public:
 	 *  Release the allocated memory, must be called regardless of the init error
 	 */
 	EXPORT void deinit();
+
+	/**
+	 *  Retrieve the mach header and __TEXT addresses for KC mode
+	 *
+	 *  @param slide load slide if calculating for kexts
+	 *  @param size  memory size
+	 *
+	 *  @return KERN_SUCCESS on success
+	 */
+	kern_return_t kcGetRunningAddresses(mach_vm_address_t slide, size_t size);
 
 	/**
 	 *  Retrieve the mach header and __TEXT addresses