+++ Changelog.md
@@ -4,6 +4,7 @@ Lilu Changelog
 - Added preliminary definitions for 11.0 support
 - Temporarily disabled user patcher for 11.0
 - Added `external-audio` property to ignore PCI audio cards
+- Added in-memory symbol solving for 11.0
 
 #### v1.4.5
 - Fixed newer CPU generation detection